+++ notebook/notes/history.md
@@ -5,6 +5,7 @@ Here's all the contests Jimmy have participated in.
 - Codeforces Round 1092 Div 2. (unrated) - 333rd 
 - Codeforces Round 1093 Div 2. - 686th - rating 622 (+622)
 - Codeforces Round 190 Div 2. - 2894th - rating 959 (+337)
+- Codeforces Round 1099 Div 2. - 5500th - rating 1160 (+201)
 
 ## AtCoder